04-orquestacion-airflow/requirements.txt
# Dependencias principales de Apache Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.0
apache-airflow-providers-redis==3.4.0
apache-airflow-providers-celery==3.4.0
apache-airflow-providers-docker==3.7.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-slack==8.0.0
apache-airflow-providers-email==1.1.0

# Base de datos y almacenamiento
postgresql==0.0.1
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Procesamiento de datos
pandas==2.1.4
numpy==1.24.4
sqlalchemy-utils==0.41.1

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
mlflow==2.8.1

# Monitoreo y métricas
prometheus-client==0.19.0
grafana-api==1.0.3

# Utilidades
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
tqdm==4.66.1

# APIs y servicios
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Notificaciones
slack-sdk==3.26.1
sendgrid==6.10.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-airflow==0.1.0

# Desarrollo
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Docker y Kubernetes
docker==6.1.3
kubernetes==28.1.0

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.29.0
ipywidgets==8.1.1

# Logging y debugging
rich==13.7.0
click==8.1.7
